// frontend/src/api/models/lineDetail.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * cryo-wiring
 * OpenAPI spec version: 0.1.0
 */
import type { LineDetailQubit } from './lineDetailQubit';
import type { LineDetailQubits } from './lineDetailQubits';
import type { LineDetailStages } from './lineDetailStages';

export interface LineDetail {
  line_id: string;
  qubit?: LineDetailQubit;
  qubits?: LineDetailQubits;
  stages: LineDetailStages;
}